requirements.txt
# streamlit
streamlit==1.39.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-drawable-canvas==0.9.3
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-js-eval==0.1.7
streamlit-keyup==0.2.4
streamlit-notifications==0.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5

# openai
openai==1.52.1

# dotenv
python-dotenv==0.19.2